backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
selenium==4.15.0
python-multipart==0.0.6
